Unsigned integer classes, but only where a division pays for them
`bytes` still lost to node after the facts fix, and node was also beating
the C++ reference, which should not happen for a loop like that. It was
signed `%`: C has to correct for the sign of the dividend, and the
unsigned form does not. Measured standalone at **1.88x**, and the facts
already prove the accumulators non-negative, so `hir::specialize` can
choose `uint32_t` for a class it can prove never holds a negative.

Three bugs, every one caught by the differential harness because its
pools contain negative numbers:

  - `i < n`. The counter's class is non-negative; `n` is a parameter and
    deliberately not joined to it, so it is not in the class and not
    consulted. `n` of -1 read as 4294967295 runs the loop four billion
    times instead of none.
  - `Math.abs`, which emits `x < 0 ? -x : x`. With `x` unsigned the test
    cannot be true, so it became the identity and `Math.abs(-32768)`
    returned 4294934528. Its *result* is legitimately non-negative, which
    is exactly why its class was eligible and its operand must not have
    been.
  - The general form of both: operands are coerced to the *operation's*
    type, so making an operation unsigned can put an unsigned cast on a
    value from outside the class entirely. Every operator that interprets
    the sign bit -- comparisons, `/`, `%`, `>>`, `abs`, `neg` -- needs
    both operands non-negative, not just its result.

Then a fourth thing, which no differential could catch because it was not
a correctness problem:

    awfy-sieve   1.74x the C++ reference -> 2.20x

`sieve` contains no division. It went unsigned because its bound is a
known 5000 and every index is non-negative, and it got 26% slower.
Signed overflow is undefined in C, and that is what lets a compiler
assume an induction variable never wraps and transform the loop on that
basis; unsigned wraparound is *defined*, so the loop has to be preserved
as written. So a class now has to contain the operation that pays.

    bytes        2.36x -> 0.97x, ahead of C++, node and bun
    awfy-sieve   unchanged

The reference moved too, and that matters for reading the number: it had
`int32_t` accumulators and zlib's `adler32` uses unsigned, so the
original comparison was against C++ I had handicapped. Against the honest
reference the win is 2.4x rather than 3.4x, and the thing being measured
is whether the compiler works out on its own what the C++ programmer
knew.

Also pins `Convert`'s transfer function with two tests -- a benchmark
told me the arm was absent, and nothing would have told me it stayed
present -- and formats four files that had gone in unformatted, `cargo
fmt --check` being enforced by nothing that runs automatically.

Record 0016 gains the postscript, and the transferable half of it:
measuring the case you are optimizing would have hidden the `sieve` cost
entirely, because `sieve` has no remainder in it and is not where anyone
would look.

+++ benches/cases/bytes/ref.cpp
@@ -1,5 +1,13 @@
 // The same checksum a C++ programmer writes: bytes in a `uint8_t` buffer, the
-// accumulators in `int32_t`.
+// accumulators in `uint32_t`.
+//
+// Unsigned because that is what zlib's `adler32` uses, and because it is the
+// faster program: C's signed remainder has to correct for the sign of the
+// dividend, which costs a multiply and two shifts more than the unsigned form.
+// Measured at 1.88x on this loop. Writing the reference with `int32_t` would
+// hand the comparison a handicap the author would not have written, and the
+// thing worth measuring is whether the compiler works out on its own what the
+// C++ programmer knew.
 //
 // The shift is done in `uint32_t` and cast back. `b << 16` with `b` above 32767
 // overflows a signed int, which is undefined behaviour in C++ and is *defined*
@@ -21,14 +29,13 @@ static std::int32_t run(std::int32_t seed) {
 
     std::int32_t total = 0;
     for (std::int32_t pass = 0; pass < 64; pass++) {
-        std::int32_t a = 1;
-        std::int32_t b = 0;
+        std::uint32_t a = 1;
+        std::uint32_t b = 0;
         for (std::int32_t i = 0; i < length; i++) {
             a = (a + data[i]) % 65521;
             b = (b + a) % 65521;
         }
-        total = total + static_cast<std::int32_t>(
-            (static_cast<std::uint32_t>(b) << 16) | static_cast<std::uint32_t>(a));
+        total = total + static_cast<std::int32_t>((b << 16) | a);
     }
     return total;
 }
